requirements.txt
astroid==2.2.5
Click==7.0
dukpy==0.2.2
Flask==1.1.1
Flask-FlatPages==0.7.1
Frozen-Flask==0.15
htmlmin==0.1.12
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
libsass==0.19.2
livereload==2.6.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
ply==3.11
pylint==2.3.1
PyYAML==5.1.2
six==1.12.0
slimit==0.8.1
tornado==6.0.3
typed-ast==1.4.0
Werkzeug==0.15.5
wrapt==1.11.2
